Change PCI MEM Space from 0x10000000~0x17ffffff to  0x40000000~0x7fffffff to support  large video frame buffer up to 512M for bridge chip connected with HT bus.
It PASS test only in GUP_SP mode, which is the defalut mode in 3A780e branch of PMON.
Address space between 0x40000000 and 0x7fffffff is map to 1G memory space before, now
system momery size is at leaset 2G, so the space is left free to be used as PCI MEM space.
While PCI IO Space and HT Space is the same as before.

NOTICE: pci mem space base address need to be changed to 0x40000000 in kenrel source code
if this patch is applied, otherwise kernel will boot fail. If the bridge chip is connected
to PCIX controler in Loongson CPU, this patch should be changed to be applied.

Target: Bontio3a780e