Fix bugs that Loader Default cannot clear the Admin Password
When you enter the PMON Console status and input "main" in the command line pattern, you
will see the window of "LOONGSON BIOS SETUP V2.1". In the label of "Safe", you can setup
admin password according to the tips, then choose Save and Exit in the label of "Exit".
Therefore, you must enter the password in the next reboot.  If you choose Loader Default,
you could clear the password.