Add beep during boot the system.
Target: 3A server
Before: When you power on the computer or detect there is no memory bank in the memory
slots, there is no sound at all.
Now: When you power on the computer or detect there is no memory bank in the memory slots,
you will listen the sound of beep.
Test method: Listen during boot the system or power on without memory bank.