Fix the error of "boot failed" after the new kernel loaded
Before: In the "readsector" function, "res = (long long)(FSC - > PartitionStart) * SECTORSIZE +
SECTORSIZE * sector", the type is not consistent, when "sector" is large, the result is not correct,
which lead to some errors.
Now: Forced convert all types to the same type, the "readsector" function is OK.
Test: Put a newly compiled kernel into a FAT formated of usb disk, boot into the pmon interface,
user can load the kernel successlly.

Targets: all