Fix the bug that we can not set the time of BIOS window by USB keyboard
Before: when we return to PMON from BIOS window, the variable "bios_available" is zero, so we can
not set the time of BIOS window by USB keyboard. It causes the problem as Bug 141 described as
http://10.2.5.24/dev/bugzilla/show_bug.cgi?id=141 showes.
Now: In the "cmd_main" function, when return to PMON, it calls "biso_available = 1", now USB keyboard
 is OK.
Test: Boot into PMON interface, when enter "main" command again, we can set the time of BIOS window by
USB keyboard.

Targets: Bonito3a780e