Fix the bug that the type of memory is error
Before: when display the size of memory, the type of memory is fixed for DDRII. It causes the
problem as Bug 195 described as http://10.2.5.24/dev/bugzilla/show_bug.cgi?id=195 showes.
Now: Removed the "DDRII", now the bios window can not display the type of memory.
Test: Boot into the pmon interface, user enter "main", the bios window can not display the type of
memory.

Targets: Bonito3a780e