Transfers poweroff/reboot from kernel to bios for 3A/3B.
Before: The code of kernel should be modified corresponding to computer of different
type.
Now: The code of kernel neednot be modified if we change the type of computer. The
realizing method is reserving low memory space to complete the function that transfers
poweroff/reboot from kernel to bios for 3A/3B.(BIOS PART) When we input "poweroff" in
the kernel, it will jump to the bios to execute the command.
Test method: Input the command of "poweroff" or "reboot" in the kernel, it will shut
down or restart.
Target: All