Fix the bug that the character "k" and the character "l" can not be uppercase
Before: In the "caps_map" variable, the characters "k" and the characters "l" were
defined as the "0xfb6b" and "0xfb6c". It caused the problem as Bug 221 described as
http://www.loongson.cn/dev/bugzilla/show_bug.cgi?id=221 showes.
Now: In the "caps_map" variable, the characters "k" and the characters "l" were
defined as the "0xfb4b" and "0xfb4c". Now in Caps-lock mode, the character "k" and the character "l"
can be uppercase.
Test: Start to pmon state, in Caps-lock mode, you observe whether the character "k" and the character
"l" can be uppercase.

Targets: all