Fix the bug that we can not delete the characters which we entered in PMON window
Before: In the "w_password" function, when we use a USB keyboard, the sentence of "if(charin==127 || charin==7)"
is wrong and can not deal with the deletion. It caused the problem
as Bug 246 described as http://www.loongson.cn/dev/bugzilla/show_bug.cgi?id=246 showes.
Now: Change the sentence of "if(charin==127 || charin==7)" to the sentence of "if(charin==127 || charin ==
'\b')", now the USB keyboard is OK.
Test: Start to pmon state, when you setup admin passwork, you observe whether we can delete the characters which
we entered.