Fix the bug that we can not get right information of net
Before: When we run "netinfo" on Bonito3a780e, the result we get is the information of
Bonito3aserver. Because the code send wrong commands to the "do_cmd" founction, so it caused the
problems as bug 247 and 248 described as http://10.2.5.24/dev/bugzilla/show_bug.cgi?id=247 and
http://10.2.5.24/dev/bugzilla/show_bug.cgi?id=248 shows.
Now: We add a judge to determine plate of the board by mocro, so we can send match
commands.
Test: Boot to PMON interface, run "info" at first, next select "netinfo" option, and run
for many times, we still get right informations.

Targets: all