Fix the problem of sata ports scanning
Before: When port1 and port2 had hard disks, we found that port2 was wd0 and port1 was
wd1,because the code scaned port2 earlier than port1, so it caused the problem as bug  278
described as http://10.2.5.24/dev/bugzilla/show_bug.cgi?id=278 shows.
Now: When port1 and port2 have hard disks, we exchange the names of their device structure in
linked list of device.
Test: Make sure that port1 and port2 have hard disks, and boot to PMON interface, run "load
/dev/fs/ext2@wd0/boot/" and "load /dev/fs/ext2@wd1/boot/", you will find that informations
are right.

Targets: All