Fix the bug that usb disk conflict with usb cd-rom
Before: When we choose usb disk or usb cd-rom to install system, the code assume that the default device is usb0, but the actual device may not be usb0, so it caused the problem asbug 182 described as  http://www.loongson.org/dev/bugzilla/show_bug.cgi?id=182 showes.
Now: When looking for configuration file failed, the code would continue to looking for itby plusing the device number for four times until find it.
Test: Insert usb disk and usb cd-rom before power-up, then startup and you will observe
that it could install system successfully by using either way.

Target:ALL