Fix the bug that shuangfeiyang usb keyboard can not work
Before: If we press the shuangfeiyan usb keyboard when the usb ohci
controller enumerate the usb devices, the usb keyboard could not work properly.
because the original code did not deal with the message from usb keyboard, so
it caused the problem as bug 382 described
http://10.2.5.24/dev/bugzilla/show_bug.cgi?id=382  shows.
Now: I add the handling code into the delay founction, so the message from the
usb keyboard can be dealed with immediately.
Test: Insert the shuangfeiyan usb keyboard, start up and press the keyboard before
the PMON command line, the keyboard can work properly.

Targets: All