Fix "make tgt=ram" compile error and speed up debug
Oirginal script file use "mipsel-linux-objdump" to dump "pmon",
but the command maybe not available on all machines for most case,
which leads failure to resolve important symbals such as "_edata"
and "_end", so "objdump" is used instread of "mipsel-linux-objdump".
Now compile error is disappeared, and "gzram" can be used to debug
pmon code in C environment, this will save time in ddr, cache and
scache  initialization.

Target: all