Modify the PCIE auto-configure code from v0.2 to v0.3.
The current version uses PCIE_*_PRSNTnX to decide which PCIE Port will be initialized,
but this not work when you put some PCIE device card into a narrower PCIE socket(like a x16 video card
in a x8 socket), because some of these cards only connect the farest PRSNTn PIN to gnd, and causes
our mother board cannot detect the device.

In this version, we use link state of each PCIE port to decide whether to enable it.
So the PCIE_*_PRSNTn0 PIN is not cared any more as declaimed in early datasheet.
As long as PCIE_*_PRSNTn1/2/3 PINs are correctly dealed on mother board, you need do nothing with the BIOS.

But you need to modify the BIOS configure code in one situation where:
you split a PCIE interface, and route Port1/2/3 to a socket, and put a wider device card into it(which should be rare).
Because in this situation, the PRSNTn1/2/3 PIN is not pull-down which cause the LS7A1000 PCIE works in solo Port
mode, so the device will never be detected.

Change-Id: I7aeafea849c799a4fa0a81ce023bd8144c922eda